Implemented parallel particle drift
The script 'skeletor.py' initializes a random particle distribution on a
grid. This grid is divided into a number of subdomains, one for each MPI
process. The particles are distributed in chunks to each process
depending on which subdomain they are located in. They are then pushed
(x += vx*dt, y += vy*dt) and, if they've left the current subdomain, are
communicated to the processor that corresponds to the subdomain they're
supposed to be in. In the end, the particles from all processors are
collected into a global particle array and the result is compared with
what one gets if the global particle array is pushed directly without
any MPI communication.