Revert "Modify electric field inside push"
This reverts commit af0adc7ef3b990ab4a16fbb185b7ef63ecaae327.

I've concluded that it is in fact not necessary
to do this. It is possible to calculate E_star
and pass it to the push instead. This does not
work for NGP interpolation but we will always be
using at least CIC.

One has to be careful with the
boundaries on the electric field, E_star.
See next commit