Many epicycles in MPI-parallized shearing box
I think we have an issue with round off error
on the boundaries - reminiscent of the one we
saw in the ppcode.

We still need to fix the add-guards function
to take into account the shear when calculating
the charge density. Right now I have removed a
line of values in the plotting of rho.