Added new example: deposition with shear
This is simply tests/test_sheared_burgers run at a single time. I added
this because the shearing periodic boundary condition looks particularly
bad at this time. Is this really as good as it gets with this way of
depositing charge and current in the shearing sheet (i.e. using FFTs to
interpolate)?