New keyword `global_init` for `InitialCondition`
When this is set to true, the particles are initialized globally (i.e.
as if the simulation were to be run on a single processor) and then
distributed to the individiual processors according to their position.
This ensures that the particle distribution is independent of the number
of processors, even if the particles' positions are randomized. The
downside is much higher memory consumption, because each processor
allocates a particle array that holds the total number of particles
across *all* processors, not just the ones in its subdomain.